+++ Sources/ZcashLightClientKit/Synchronizer/SDKSynchronizer.swift
@@ -12,7 +12,6 @@ import Combine
 /// Synchronizer implementation for UIKit and iOS 13+
 // swiftlint:disable type_body_length
 public class SDKSynchronizer: Synchronizer {
-    
     public var alias: ZcashSynchronizerAlias { initializer.alias }
 
     private lazy var streamsUpdateQueue = { DispatchQueue(label: "streamsUpdateQueue_\(initializer.alias.description)") }()
@@ -300,9 +299,9 @@ public class SDKSynchronizer: Synchronizer {
         do {
             try throwIfUnprepared()
             return try await transactionEncoder.proposeFulfillingPaymentFromURI(
-               uri,
-               accountIndex: accountIndex
-           )
+                uri,
+                accountIndex: accountIndex
+            )
         } catch ZcashError.rustCreateToAddress(let error) {
             throw ZcashError.rustProposeTransferFromURI(error)
         } catch {