+++ Tests/OfflineTests/ClosureSynchronizerOfflineTests.swift
@@ -685,12 +685,14 @@ class ClosureSynchronizerOfflineTests: XCTestCase {
                 PoolBalance(
                     spendableValue: Zatoshi(333),
                     changePendingConfirmation: .zero,
-                    valuePendingSpendability: .zero),
+                    valuePendingSpendability: .zero
+                ),
             orchardBalance:
                 PoolBalance(
                     spendableValue: Zatoshi(333),
                     changePendingConfirmation: .zero,
-                    valuePendingSpendability: .zero),
+                    valuePendingSpendability: .zero
+                ),
             unshielded: .zero
         )
         
@@ -739,12 +741,14 @@ class ClosureSynchronizerOfflineTests: XCTestCase {
                 PoolBalance(
                     spendableValue: .zero,
                     changePendingConfirmation: Zatoshi(333),
-                    valuePendingSpendability: .zero),
+                    valuePendingSpendability: .zero
+                ),
             orchardBalance:
                 PoolBalance(
                     spendableValue: .zero,
                     changePendingConfirmation: Zatoshi(333),
-                    valuePendingSpendability: .zero),
+                    valuePendingSpendability: .zero
+                ),
             unshielded: .zero
         )